chore(emails): fix lint/type/i18n issues from rebrand verification
- fix mypy: use list(img.getdata()) to avoid ImagingCore not-iterable error
- fix i18n: run makemessages, defuzz 14 entries, add missing payout-doc translation (de/fr/it), fix %(currency)s in msgstr for 4 strings whose msgids no longer include that placeholder
- split test_ticket_templates.py at 1038 lines → extract TestTicketPdfBranding and TestTicketCancelledStaffTemplateBranching into test_ticket_templates_branding.py (now both files under 1000-line limit)
- ruff auto-fixed 5 formatting issues and 1 lint error across rebrand test files
